PoC for mixed dynamic rule indicator in popup matrix
This is a quick demonstration of how inconsistencies
caused by conflicting dynamic block/allow rules in the
dynamic blocking matrix could be handled. A yellow-ish
indicator is added to cells that are affected by more
than one conflicting dynamic filtering rule.

For example if 3p deny and 3p-scripts or 3p-frames allow/noop
are selected, the current way of shading 3p cells red/gray
does not accurately reflect the mixed state for those cells.
For first-party cells, inline-script and 1p-script are affected.
The display of image and all rules hasn't been changed.

This is based on the idea outlined in my comment
https://github.com/gorhill/uBlock/issues/527#issuecomment-124918048
refer to that for screenshots and further examples.

https://github.com/gorhill/uBlock/issues/544